requirements.txt
gradio==4.44.1
python-dotenv==1.0.0
numpy==1.24.4
